Remove unused monolithic route.ts files from lazy-loaded pages
Deleted 4 unused monolithic API route files that were replaced by split endpoints during lazy loading implementation:
- page-flow-analysis/route.ts
- returning-analysis/route.ts
- environment-analysis/route.ts
- campaign-analysis/route.ts

All pages now use split endpoints (e.g., /insights, /utm, /metrics) for independent section loading. These monolithic files are dead code and no longer referenced.
